[FEAT] 테그 필터링 제외 다 구현
