+++ People_zero/urls.py
@@ -15,5 +15,6 @@ urlpatterns = [
     path('api-auth/', include('rest_framework.urls', namespace='rest_framework')),
     path('test1/', include('accounts.urls')),
     path('storetest/', include('store.urls')),
+    path('post/',include('board.urls'))
 ]+ static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
