+++ board/models.py
@@ -20,15 +20,8 @@ class Post(TimestampAbstractModel):
     recommend_cnt = models.IntegerField(default=0)
     tag_set = models.ManyToManyField("Tag", blank=True)
    
-    #FIXME 일단 content에서 (POST객체에서 분리) 테그를 분리하여 list로 만들어주는
-    #함수입니다.. 왜 안될까..마음이 너무나 아파버려...
-    def extract_tag_list(self):
-        tag_name_list = re.findall(r"#([a-zA-Z\dㄱ-힣]+)",self.content)
-        tag_list = []
-        for tag_name in tag_name_list:
-            tag,_ = models.Tag.objects.get_or_create(name=tag_name)
-            tag_list.append(tag)
-        return tag_list
+    def __str__(self):
+        return self.title
 
     class Meta:
         ordering = ["-id"] # 최신순 정렬
@@ -39,10 +32,16 @@ class Post(TimestampAbstractModel):
 class Tag(TimestampAbstractModel):
     name = models.CharField(max_length = 30, primary_key=True)
 
+    def __str__(self):
+        return self.name
+
 class Comment(TimestampAbstractModel):
     # user = models.ForeignKey(User, on_delete=models.SET_NULL, null=True)
     post = models.ForeignKey(Post, on_delete=models.CASCADE)
     content = models.TextField()
+
+    def __str__(self):
+        return self.content
 
     class Meta:
         ordering = ["-id"]
